new Update

+++ ft_puthex_lower.c
@@ -1,6 +1,6 @@
-#include "printf.h"
+#include "ft_printf.h"
 
-void ft_puthex_lower(unsigned long n, int *ret)
+void ft_puthex_lower(unsigned int n, int *ret)
 {
 	char	*base;
